spill main project/spilleu/requirements.txt
Django==5.2.1
djangorestframework
djangorestframework-simplejwt
django-cors-headers
python-decouple
dj-database-url
psycopg2-binary
cloudinary
django-cloudinary-storage
Pillow
gunicorn
whitenoise
exponent-server-sdk